+++ ColorPicker.Core.Tests/Utilities/ColorConversionsTests.cs
@@ -11,11 +11,9 @@ namespace ColorPicker.Tests.Utilities
         [InlineData(255, 255, 255, 0, 0, 1)]
         [InlineData(0, 0, 0, 0, 0, 0)]
         [InlineData(128, 128, 128, 0, 0, 0.502)]
-        public void RgbToHsv_PrimaryColors(
-            byte r, byte g, byte b,
-            double eh, double es, double ev)
+        public void RgbToHsv_PrimaryColors(byte r, byte g, byte b, float eh, float es, float ev)
         {
-            ColorConversions.RgbToHsv(r, g, b, out double h, out double s, out double v);
+            ColorConversions.RgbToHsv(r, g, b, out float h, out float s, out float v);
 
             Assert.Equal(eh, h, 3);
             Assert.Equal(es, s, 3);
@@ -26,9 +24,7 @@ namespace ColorPicker.Tests.Utilities
         [InlineData(0, 1, 1, 255, 0, 0)]
         [InlineData(120, 1, 1, 0, 255, 0)]
         [InlineData(240, 1, 1, 0, 0, 255)]
-        public void HsvToRgb_PrimaryColors(
-            double h, double s, double v,
-            byte er, byte eg, byte eb)
+        public void HsvToRgb_PrimaryColors(float h, float s, float v, byte er, byte eg, byte eb)
         {
             ColorConversions.HsvToRgb(h, s, v, out byte r, out byte g, out byte b);
 
@@ -43,7 +39,7 @@ namespace ColorPicker.Tests.Utilities
         [InlineData(0, 128, 255)]
         public void RgbToHsvBackToRgb_ReturnsOriginal(byte r, byte g, byte b)
         {
-            ColorConversions.RgbToHsv(r, g, b, out double h, out double s, out double v);
+            ColorConversions.RgbToHsv(r, g, b, out float h, out float s, out float v);
             ColorConversions.HsvToRgb(h, s, v, out byte rr, out byte gg, out byte bb);
 
             Assert.InRange(rr, r - 1, r + 1);
@@ -55,11 +51,9 @@ namespace ColorPicker.Tests.Utilities
         [InlineData(255, 0, 0, 0, 1, 0.5)]
         [InlineData(0, 255, 0, 120, 1, 0.5)]
         [InlineData(0, 0, 255, 240, 1, 0.5)]
-        public void RgbToHsl_PrimaryColors(
-            byte r, byte g, byte b,
-            double eh, double es, double el)
+        public void RgbToHsl_PrimaryColors(byte r, byte g, byte b, float eh, float es, float el)
         {
-            ColorConversions.RgbToHsl(r, g, b, out double h, out double s, out double l);
+            ColorConversions.RgbToHsl(r, g, b, out float h, out float s, out float l);
 
             Assert.Equal(eh, h, 3);
             Assert.Equal(es, s, 3);
@@ -70,11 +64,11 @@ namespace ColorPicker.Tests.Utilities
         [InlineData(19.84, 0.5, 0.5)]
         [InlineData(200.16, 1.0, 0.25)]
         [InlineData(300, 0.29, 0.8)]
-        public void HslToRgbRoundTrip(double h, double s, double l)
+        public void HslToRgbRoundTrip(float h, float s, float l)
         {
             ColorConversions.HslToRgb(h, s, l, out byte r, out byte g, out byte b);
 
-            ColorConversions.RgbToHsl(r, g, b, out double hh, out double ss, out double ll);
+            ColorConversions.RgbToHsl(r, g, b, out float hh, out float ss, out float ll);
 
             Assert.Equal(h, hh, 2);
             Assert.Equal(s, ss, 2);
@@ -84,7 +78,7 @@ namespace ColorPicker.Tests.Utilities
         [Fact]
         public void HslToRgb_ZeroSaturation_ProducesGray()
         {
-            ColorConversions.HslToRgb(123, 0, 0.6, out byte r, out byte g, out byte b);
+            ColorConversions.HslToRgb(123, 0, 0.6f, out byte r, out byte g, out byte b);
 
             Assert.Equal(r, g);
             Assert.Equal(g, b);
